requirement.txt
I2C-HD77480-LCD >= 0.1.1
jproperties~=2.1.1
setuptools~=60.2.0
